Breaking change: Upgraded to 64 bit MessageIDs as requested and required by the PCI version of the InterFAX API.

+++ InterFAX.Api.Test.Integration.Core/DocumentsTests.cs
@@ -143,7 +143,7 @@ namespace InterFAX.Api.Test.Integration
         [TestMethod]
         public void can_fax_small_document_as_stream()
         {
-            int faxId;
+            Int64 faxId;
             using (var fileStream = File.OpenRead(Path.Combine(_testPath, "test.pdf")))
             {
                 // Fax the document
